chore: add discovery tests for branch
