+++ Source/GitSnapshotter.UnitTests/GitSnapshotter.UnitTests.csproj
@@ -9,6 +9,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <PackageReference Include="AutoFixture.Xunit2" Version="4.18.1" />
         <PackageReference Include="CliWrap" Version="3.7.0" />
         <PackageReference Include="coverlet.collector" Version="6.0.2" />
         <PackageReference Include="FluentAssertions" Version="7.0.0" />
@@ -24,6 +25,7 @@
 
     <ItemGroup>
         <Using Include="Xunit" />
+        <Using Include="AutoFixture.Xunit2" />
     </ItemGroup>
 
     <ItemGroup>
